Extract codeSelection strategy
